chore: Extracted type `CurrentUser` into the base type definition file
