+++ src/main/webapp/resources/js/pages/search/loaders/user-loader.ts
@@ -1,4 +1,9 @@
-import { CurrentUser, fetchCurrentUserDetails } from "../../../apis/users/user";
+import { fetchCurrentUserDetails } from "../../../apis/users/user";
+import type { CurrentUser } from "../../../types/irida";
+
+/**
+ * Loader for the react-router for the current user
+ */
 export default async function loader(): Promise<CurrentUser> {
   return await fetchCurrentUserDetails();
 }
